+++ editor/imgui/gists/imgui_bezier.cpp
@@ -116,7 +116,7 @@ namespace ImGui
                 for (int i = 0; i < IM_ARRAYSIZE(presets); ++i) {
                     if( i == 1 || i == 9 || i == 17 ) ImGui::Separator();
                     if (ImGui::MenuItem(presets[i].name, NULL, P[4] == i)) {
-                        P[4] = i;
+                        P[4] = (float)i;
                         reload = 1;
                     }
                 }
@@ -164,13 +164,13 @@ namespace ImGui
         RenderFrame(bb.Min, bb.Max, GetColorU32(ImGuiCol_FrameBg, 1), true, Style.FrameRounding);
 
         // background grid
-        for (int i = 0; i <= Canvas.x; i += (Canvas.x / 4)) {
+        for (int i = 0; i <= Canvas.x; i += (int)(Canvas.x / 4)) {
             DrawList->AddLine(
                 ImVec2(bb.Min.x + i, bb.Min.y),
                 ImVec2(bb.Min.x + i, bb.Max.y),
                 GetColorU32(ImGuiCol_TextDisabled));
         }
-        for (int i = 0; i <= Canvas.y; i += (Canvas.y / 4)) {
+        for (int i = 0; i <= Canvas.y; i += (int)(Canvas.y / 4)) {
             DrawList->AddLine(
                 ImVec2(bb.Min.x, bb.Min.y + i),
                 ImVec2(bb.Max.x, bb.Min.y + i),
